+++ CHANGELOG.md
@@ -33,3 +33,4 @@
 * Use promoted constructors of `Capacity` instead of separated datatypes
 * Use `MultiParamTypeClasses` instead of `FlexibleIntances` for `Endpoint`
 * Use `UndecidableInstances` and `UndecidableSuperClasses` to use `Object` type family
+* Replace `Access` module with `Accessible` `Property`